feat: implement custom transformer
To avoid having a whole scikit-learn library as a dependency on a
server. The only thing needed from whole library is just to apply and
deapply function, that's it. This could be a simple custom wrapper
instead of having a whole library to it.

As a result, no scikit-learn library is needed on clinet/server.