feat: let the model do prediction in minutes
To keep seconds in game, they can be just configured, but default is in
minutes. Through trial and error, it seems like minutes have better
accuracy than seconds. Probably because a lot of builds have the minutes
in low range and to predict minutes for them instead of seconds means
less error (in absolute numbers), so the model can focus rather on more
demanding builds instead. Another fact is that RMSE punishes bigger
mistakes more, making it even more to focus on more demanding builds.